requirements.txt
matplotlib
microsoft-bonsai-api
python-dotenv
simpy
